requirements.txt
streamlit
yfinance
requests
python-dotenv
